pyproject.toml
[tool.poetry]
name = "checkcert"
version = "0.3.0"
description = "CLI to check tls cert information and determine validity"
authors = ["Alex Kelly <kellya@arachnitech.com>"]
readme = "README.md"
homepage = "https://github.com/kellya/checkcert"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
pyOpenSSL = "^21.0.0"

[tool.poetry.dev-dependencies]
python-semantic-release = "^7.19.2"
coverage = "^5.5"
pylint = "^2.11.1"
pyflakes = "^2.3.1"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
checkcert = 'checkcert.checkcert:main'

[tool.semantic_release]
version_variable = [
    "checkcert/checkcert.py:__version__",
    "pyproject.toml:version"
]
